+++ src/pages/archive.js
@@ -2,11 +2,7 @@ import React from "react";
 import TaskScreen from "../screens/TaskScreen";
 
 const Archive = () => {
-  return (
-    <>
-      <TaskScreen />
-    </>
-  );
+  return <TaskScreen />;
 };
 
 export default Archive;
