+++ src/providers/TodayContext.js
@@ -1,0 +1,9 @@
+import React, { useContext } from "react";
+
+const TodayContext = () => {
+  const { tasks } = useContext(TaskContext);
+
+  return <div></div>;
+};
+
+export default TodayContext;
